[cli] Log Secret Decryption events when using 3rd party secrets + service
For users who use the Pulumi Service for their state BUT a 3rd party
secrets provider, we now log when that secret decryption has happened.

We have 2 forms of logs:

* single secret when a user uses `pulumi config get x --show-secrets`
* batch decryption for when a user runs a stack export or config list

The API logs to request are fire and forget. The API response is a 204
so we don't need to do anything with the response.